refactor: ♻️ helpモードの状態管理をFSD依存のものに変更
