+++ packages/app/src/features/help/help.store.ts
@@ -1,0 +1,3 @@
+import { atom } from "jotai";
+
+export const helpModeAtom = atom<boolean>(false);
